Support TrueSkill ratings for 1v1 matches.
This patch adds primitive support for TrueSkill, creating a
TrueSkillPlayer and a TrueSkillEngine that supports match! for 1v1,
non-draw matches. This represents limited support for TrueSkill as a
whole, which allows for N-team matches and explicitly models draws.

Support for TrueSkill is found in relative.trueskill, and meets the
standard relative rating protocols.
Begin work on TrueSkill implementation.

mu-additive-factors is seems to be currently broken.

Less broken, but still broken, implementation

Make variable names not super shitty

add primitive support for 1v1 trueskill

This commit adds a basic implementation of trueskill for 1v1 games where a draw
is impossible.